Move thin_client RPC requests into rpc_request; de-mut thin_client
